Revert structure to single application
