--- source/apps/shared/block.tsx
@@ -1,18 +1,0 @@
-import debug from 'debug';
-import React from 'react';
-import NoSSR from 'react-no-ssr';
-import { environment } from '../../environment';
-import { Footer, Header, Layout } from '../../widgets/layout';
-
-if (environment.DEBUG) {
-  debug.enable(environment.DEBUG);
-}
-
-export const BlockPage = () => (
-  <NoSSR>
-    <Layout>
-      <Header />
-      <Footer />
-    </Layout>
-  </NoSSR>
-);
