--- source/apps/shared/no-search-results.tsx
@@ -1,20 +1,0 @@
-import debug from 'debug';
-import React from 'react';
-import NoSSR from 'react-no-ssr';
-import { environment } from '../../environment';
-import NoSearchResult from '../../features/search/components/NoSearchResult';
-import { Footer, Header, Layout } from '../../widgets/layout';
-
-if (environment.DEBUG) {
-  debug.enable(environment.DEBUG);
-}
-
-export const NoSearchResultsPage = () => (
-  <NoSSR>
-    <Layout hasContainer>
-      <Header />
-      <NoSearchResult />
-      <Footer />
-    </Layout>
-  </NoSSR>
-);
