--- source/apps/shelley/index.scss
@@ -1,53 +1,0 @@
-.headerBackgroundAnimationContainer {
-  display: flex;
-  justify-content: center;
-  left: 0;
-  position: absolute;
-  top: 0;
-  width: 100%;
-
-  .headerBackgroundAnimation {
-    height: 415px;
-    width: 1099px;
-
-    .animationIframe {
-      border: 0;
-      height: 100%;
-      opacity: 0.3;
-      width: 100%;
-    }
-  }
-}
-
-.sideBackgroundImageContainer {
-  align-items: flex-end;
-  bottom: 0;
-  display: flex;
-  position: absolute;
-  right: 0;
-  z-index: -1;
-
-  .sideBackgroundImage {
-    height: auto;
-    object-fit: contain;
-    opacity: 0.3;
-    width: 596.5px;
-  }
-}
-
-.epochList {
-  margin-top: 71px;
-  margin-bottom: 21px;
-
-  @media (max-width: 575px) {
-    margin-top: 51px;
-  }
-}
-
-.blockList {
-  margin-bottom: 139.5px;
-
-  @media (max-width: 575px) {
-    margin-bottom: 83px;
-  }
-}
