+++ source/pages/outdated-browser.tsx
@@ -1,10 +1,22 @@
 import React from 'react';
 import NoSSR from 'react-no-ssr';
 import { environment } from '../environment';
+import OutdatedBrowser from '../features/outdated-browser/OutdatedBrowser';
+import { Footer, Layout } from '../widgets/layout';
 
-let Page = () => <NoSSR />;
+let OutdatedBrowserPage = () => <NoSSR />;
 if (environment.IS_CLIENT) {
-  Page = require('../apps/shared/outdated-browser').OutdatedBrowserPage;
+  OutdatedBrowserPage = () => (
+    <NoSSR>
+      <Layout hasContainer>
+        <OutdatedBrowser
+          text="Your web browser is out of date"
+          updateBrowserUrl="http://browser-update.org/update.html?force_outdated=true"
+        />
+        <Footer />
+      </Layout>
+    </NoSSR>
+  );
 }
 
-export default Page;
+export default OutdatedBrowserPage;
